Added a 3rd party (public domain) crypto implementation (many source files) Created a udCrypto interface/module that includes specific source files for implementations, currently supporting AES256 and SHA1 - to be extended as necessary
git-svn-id: https://bne-fs-pf-001.euclideon.local/svn/Dev/trunk/src@731 f54097bf-93f5-7042-b87d-fd266665e7ff

+++ udPlatform/Source/udFile.cpp
@@ -72,7 +72,7 @@ udResult udFile_Load(const char *pFilename, void **ppMemory, int64_t *pFileLengt
       if (result != udR_Success)
         goto epilogue;
     }
-    UDASSERT(length >= alreadyRead, "Logic error in read loop");
+    UDASSERT((size_t)length >= alreadyRead, "Logic error in read loop");
     if ((size_t)length != alreadyRead)
     {
       length = alreadyRead;
